# Increased performance for the target distribution view.