--- a/app/views/backend/target_distributions/edit_many.html.haml
+++ b/app/views/backend/target_distributions/edit_many.html.haml
@@ -1,33 +1,35 @@
-- Ekylibre.profile(:distributions_2) do
-  = form_tag backend_target_distributions_path, method: :patch do
-    = hidden_field_tag(:redirect, params[:redirect]) if params[:redirect]
-    = hidden_field_tag(:activity_id, params[:activity_id]) if params[:activity_id]
-    = field_set do
-      %table
-        %thead
+= form_tag backend_target_distributions_path, method: :patch do
+  = hidden_field_tag(:redirect, params[:redirect]) if params[:redirect]
+  = hidden_field_tag(:activity_id, params[:activity_id]) if params[:activity_id]
+  = field_set do
+    %table
+      %thead
+        %tr
+          %th= TargetDistribution.human_attribute_name(:target)
+          %th= TargetDistribution.human_attribute_name(:activity_production)
+      %tbody
+        :ruby
+          activity = Activity.find_by(id: params[:activity_id]) if params[:activity_id]
+
+          product_ids = @target_distributions.map(&:target_id).compact.uniq
+          products = Product.where(id: product_ids)
+          products = product_ids.zip(products).to_h
+
+          activity_productions = ActivityProduction.all
+          activity_productions = activity_productions.of_activity(activity) if activity
+          activity_productions = activity_productions.includes(:activity, :cultivable_zone).map { |ac| [ac.name, ac.id] }
+
+        - @target_distributions.each do |distribution|
+          - target = products[distribution.target_id]
           %tr
-            %th= TargetDistribution.human_attribute_name(:target)
-            %th= TargetDistribution.human_attribute_name(:activity_production)
-        %tbody
-          :ruby
-            activity = Activity.find_by(id: params[:activity_id]) if params[:activity_id]
-
-            product_ids = @target_distributions.map(&:target_id).compact.uniq
-            products = Product.where(id: product_ids)
-            products = product_ids.zip(products).to_h
-
-          - @target_distributions.each do |distribution|
-            - target = products[distribution.target_id]
-            %tr
-              %td{style: "width: 50%"}
-                = link_to target.name, backend_product_path(target)
-                - name = "target_distributions[#{distribution.id}][target_id]"
-                -# = f.input :target_id, as: :hidden, value: target.id
-                = hidden_field_tag name, target.id
-              %td{style: "width: 50%"}
-                - activity_scope = { of_activity: activity.id } if activity
-                = selector_tag :"target_distributions[#{distribution.id}][activity_production_id]", { controller: 'backend/activity_productions', action: :unroll, scope: activity_scope }, { value: distribution.activity_production_id }
-                -# = f.association :activity_production, wrapper: :nested, collection: activity_productions, include_blank: true
-    = form_actions do
-      = submit_tag(:update.tl, data: {disable_with: :please_wait.tl}, class: 'btn btn-primary')
-      = link_to(:cancel.tl, (params[:redirect] || { controller: :activities, action: :index }), (params[:dialog] ? { class: 'btn', data: {close_dialog: params[:dialog]} } : { class: 'btn' }))
+            %td
+              = link_to target.name, backend_product_path(target)
+              - name = "target_distributions[#{distribution.id}][target_id]"
+              = hidden_field_tag name, target.id
+            %td
+              - activity_scope = { of_activity: activity.id } if activity
+              = select_tag :"target_distributions[#{distribution.id}][activity_production_id]", options_for_select(activity_productions, distribution.activity_production_id), include_blank: true
+              -# = selector_tag :"target_distributions[#{distribution.id}][activity_production_id]", { controller: 'backend/activity_productions', action: :unroll, scope: activity_scope }, { value: distribution.activity_production_id }
+  = form_actions do
+    = submit_tag(:update.tl, data: {disable_with: :please_wait.tl}, class: 'btn btn-primary')
+    = link_to(:cancel.tl, (params[:redirect] || { controller: :activities, action: :index }), (params[:dialog] ? { class: 'btn', data: {close_dialog: params[:dialog]} } : { class: 'btn' }))