Heavily rewrote view and controller action to increase performance.

On our largest local dataset (About 1040 items to compute for and render) we have a ~20x increase in performance with a drop from ~311s (311 105ms) to ~15s (15 620ms).

It is possible to increase performance again (to about 12s (12 176ms)) by using unroll selectors (selector.js) instead of `select_tag`s  but we then have an extra 20s on the client side before the full page is loaded + a strain on the client's resources + as many requests as there are items performed to the server as the unroll selectors fill up.